+++ elements/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.24.1
+
+### Fixed
+
+- [pos-make-findable](../docs/elements/components/pos-make-findable)
+  - also emits index update event, when one of multiple indexes is chosen
+
 ## 0.24.0
 
 ### Changed